// prototype-parts-marking-development/src/client-app/src/models/api/items/package/packageCreateData.ts
export interface IPackageCreateData
{
  outletMoniker: string;
  productGroupMoniker: string;
  locationMoniker: string;
  gateLevelMoniker: string;
  partMoniker: string;
  evidenceYear: number;
  initialCount: number;
  customer: string;
  project: string;
  projectNumber: string;
  comment: string;
  ownerId: number;
}